+++ src/utils/index.ts
@@ -1,3 +1,6 @@
 export * from "./getfileContents";
+export * from "./insertContent";
+export * from "./insertLineByLine";
 export * from "./parseWinPath";
+export * from "./replaceContent";
 export * from "./sleep";